refactor out annotation shenanigans to simplify DatastoreImpl a bit
